refactor: extract common data display components
